requirements.txt
Chameleon==2.18
Mako==1.0.0
MarkupSafe==0.23
MySQL-python==1.2.5
PasteDeploy==1.5.2
PyJWT==0.4.0
PyMySQL==0.6.3
Pygments==2.0.1
SQLAlchemy==0.9.8
WebOb==1.4
argparse==1.2.1
coolstorybro==0.0
pyramid==1.5.2
pyramid-chameleon==0.3
pyramid-debugtoolbar==2.2.2
pyramid-mako==1.0.2
pyramid-tm==0.10
repoze.lru==0.6
requests==2.5.1
transaction==1.4.3
translationstring==1.3
venusian==1.0
waitress==0.8.9
wsgiref==0.1.2
zope.deprecation==4.1.1
zope.interface==4.1.2
zope.sqlalchemy==0.7.5
